Reverse the renaming of stream contents
It was mentioned in PR#7 that the renaming of these stream contents could require further XML changes. Since there is uncertainty here and there is currently no harm in keeping the generic stream names, I'm reversing the renaming for now. This should be revisited when we're confident that the name changes won't result in any unknown knock on effects.